feat/reorganize-patterns: Refactored
