+++ src/behavioral-patterns/abstract-factory.js
