+++ src/behavioral-patterns/factory-method.js
