+++ src/behavioral-patterns/prototype.js
