+++ src/behavioral-patterns/singleton.js
