+++ src/creational-patterns/chain-of-responsibility.js
