+++ src/creational-patterns/command.js
