+++ src/creational-patterns/iterator.js
