+++ src/creational-patterns/mediator.js
