+++ src/creational-patterns/memento.js
