+++ src/creational-patterns/state.js
