+++ src/creational-patterns/strategy.js
