+++ src/creational-patterns/template-method.js
