+++ src/structural-patterns/adapter.js
