+++ src/structural-patterns/composite.js
