+++ src/structural-patterns/decorator.js
