+++ src/structural-patterns/facade/facade.js
