+++ src/structural-patterns/facade/index.js
