+++ src/structural-patterns/facade/requestExtractor.js
