+++ src/structural-patterns/facade/responseBuilder.js
