+++ src/structural-patterns/facade/transformer.js
