+++ src/structural-patterns/flyweight.js
