+++ src/structural-patterns/proxy.js
